+++ crates/kira/src/manager/settings.rs
@@ -8,15 +8,15 @@ use crate::backend::Backend;
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Capacities {
 	/// The maximum number of mixer sub-tracks that can exist at a time.
-	pub sub_track_capacity: u16,
+	pub sub_track_capacity: usize,
 	/// The maximum number of mixer send tracks that can exist at a time.
-	pub send_track_capacity: u16,
+	pub send_track_capacity: usize,
 	/// The maximum number of clocks that can exist at a time.
-	pub clock_capacity: u16,
+	pub clock_capacity: usize,
 	/// The maximum number of modulators that can exist at a time.
-	pub modulator_capacity: u16,
+	pub modulator_capacity: usize,
 	/// The maximum number of listeners that can exist at a time.
-	pub listener_capacity: u16,
+	pub listener_capacity: usize,
 }
 
 impl Default for Capacities {
